feat: add OIDC login to Asgardeo and display of user/claims for test/debug
This commit adds the OIDC-related settings that mozilla_django_oidc requires
and enables login to and logout from Asgardeo.  There is a basic index page
that shows the logged-in user model and the claims that have been found
in the id token.  Note that because mozilla_djano_oidc does not have type
hints there are a few places where typing is ignored, and I've also added it
to the overrides in pyproject.toml.  This should be fixed in a more
compliant way for prod.