feat: add settings and scopes for access register your data
This commit adds the necessary scopes to settings.py that we
need to request when logging in a user via the identity service.
These are needed in order to have the correct permissions to
call the Register Your Data API.  The commit also adds the base
URL for the Register Your Data API as an env var.