refactor+feat: refactor identity service code and add error handling/logging
This commit carries out a combination of refactoring and adding error
handling and logging.  It moves identity service interactions into the
user model as this is where the fields are extracted from claims,
and the source of data when writing to the identity service.  Exception
handling is added and logging incorporated, including a (non-encrypted
for the moment) audit log.  We also refactor the account provisioning
code to improve error handling and logging.  The two view functions
that were in identity.py have been moved into views.py.  This is
generally a bit of a tidy up and refactor.