feat: added models and tests for RYD reporting org/user data
This commit adds a set of models for data coming back from the RYD API
so that we can perform error handling/parsing RYD responses in the models
rather than in view functions or directly in forms.  This will then
require changes for our Django forms to become ModelForms, so simplifying
the downstream code.