feat: added environment variables to configure RYD redirect behaviour
The RegisterYourDataSession initialiser takes an allow_redirects
parameter and uses a simplistic function to decide if authorisation
headers should be stripped from the requests during redirection.
This commit adds environment variables to control this behaviour
and enables using a more complicated function for determining
if authorisation headers should be stripped.  This commit closes
https://github.com/IATI/iati-account-web/issues/15.