fix: for non-admins, hide user role dropdowns on organisation detail pages
This commit hides user role dropdowns for non-admin users on organisation
detail pages.  This reduces confusion for end users when they try to do
this and then get an error message.  As part of this change, this commit
adds a property in the UsersAndRole model to get whether the user can
edit user roles.  This closes https://github.com/IATI/iati-account-web/issues/47