fix: adds super_admin role to UserAndRole model and updates templates
This fix adds the super_admin role to the UserAndRole model to prevent
an unhandled exception when creating a model for a record when viewed
as superadmin.  It also updates the user role text at the top of each
page to show when something is being viewed as a superadmin.